Add files via upload
Program 1: Write a program to lower all the characters in the given strings and remove all non-alphanumeric characters from the string and \
check after removing if the string is palindrome or not.  
GIVEN STRING IS : A man, A plan, A goodman

Program 2: WAP to fetch the strings in the given list into new list

Program 3: WAP to exchange first and last character of the string

Program 4 : Write a Python program to remove the nth index character from a nonempty string.

Program 5 : Write a Python program to insert a string in the middle of a string.

Program 6: Write a Python program that takes a list of words and return the longest word and the length of the longest one.

Program 7: Write a Python program to get a new string from a given string where "Is" has been added to the front. 
If the given string already begins with "Is" then return the string unchanged.(without using inbuilt function)

Program 8: Write a Python program that accepts a comma separated sequence of words as input and prints the unique words in 
sorted form (alphanumerically).

Program 9 : Wap to check whether entered pincode is valid or not valid pincode should have 6 digits

Program 10: Write a program that accepts a sentence and calculate the number of letters and digits.